refactor: split env command modules
